feat: generateIcon を作成
